\name{centRoidSE-package}
\alias{centRoidSE-package}
\alias{centRoidSE}
\docType{package}
\title{
Method to calculate standard errors and effect sizes from multiple reported values.
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab centRoidSE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-18\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Andrew W Brown
John A Dawson

Maintainer: Andrew W Brown<awbrown@uab.edu>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
